Added the new installed version build file.
Added the concept of a PackRefreshListener so that new image downloads would properly reload the pack images in the selector.

Added downloading/caching of the old forge libraries from 1.4/1.5 to the ModpackInstaller.

Updated the ModpackInstaller to not redownload/install when the modpack is installed.

Updated the ModpackInstaller to not redownload/install the completeversion and minecraft when a modpack is installed, except if the install was a migration from the legacy installed file.

Signed-off-by: Dalton Pelc <daltonjpelc@gmail.com>