Rewrote the installation process, added 1.7.4 support
This installation rewrite job-ifies most units of installation logic, and
then I add jobs for installing resources.  This adds support for the new
resources system, as well as the major layout changes added in 1.7.3

Additionally, support was added for new auth responses which are then
piped into the launch code for Minecraft.