Optimise the launcher's UI images
All of the PNGs used by the launcher for the UI were un-optimised. I
have optimised them losslessly, meaning that no quality is lost but
filesizes are reduced. With these optimised PNG files, less I/O is
required to load the images and the same or even sometimes less CPU
usage is required to decode these images.

**Thanks to these optimisations, the launcher could load up about 5-10%
faster.**

The command used to optimise the PNG images was: optipng -o7 -zm1-9 -nx
*.png