Added major UI pass
This change includes updating the login UI frame to handle both login
types more elegantly, as well as changing some of the way in which the
login code is organized. This is done to reduce the amount of places
state is tracked and performed on for the login flow.

Remaining items include:
* Further UI polish on the login frame
* Either fixing or removing the 'remember' tickbox
* Corner case testing/error handling, especially with Microsoft logins
* Solving for token duration for Microsoft accounts
* Port in use problems with oauth
* Random client token back for Mojang accounts