Adding exception and edge case handling flow
Covers a lot of changes from a day of debugging and working through auth
problems. Major updates include partial token refreshes and storing the
Xbox access token in order to reduce the frequency with which users will
need to do a browser auth.

There are bugs still in this state, especially around edge case and
exception handling, but the main user flow seems OK for valid
mojang/microsoft.

Main reason for a lot of the refactoring is to be able to safely handle
exception flow in the login frame, with auth specific code being moved
and retained in auth specific concrete instances.