Updated MinecraftLauncher to use interface for type and properties
This change moves handling the user type the minecraft launcher
expects as well as the properties json blob out from conditional logic
in the launch minecraft code, and into the user interface.

Concrete implementations of the logic now exist within the user type
implementations for both Mojang and Microsoft accounts.

This keeps dependencies on concrete implementations from bleeding out
into the rest of the codebase.