Switched user.json microsoft user type symbol
The minecraft game itself expects "msa" as the user type signifier for a
microsoft account, not "microsoft" so in order to maintain parity and
clarity with debugging and future work, this change updates the json
stored to use "msa" instead of "microsoft".

It will support loading existing users that are listed as type
"microsoft", but then save them back as "msa" after.